pyproject.toml
[project]
name = "tds-project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.2.0",
    "fastapi>=0.115.8",
    "flask>=3.1.0",
    "gitpython>=3.1.44",
    "markdown>=3.7",
    "openai==0.28",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.3",
    "speechrecognition>=3.14.1",
    "uvicorn>=0.34.0",
]
